Text_Collector/requirements.txt
pandas>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
urllib3>=1.26.0
python-dotenv>=0.15.0